+++ src/containers/Company/SalaryWorkTimeScreen.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import TimeAndSalaryScreen from '../../components/Company/TimeAndSalaryScreen';
+import SalaryWorkTimeScreen from '../../components/Company/SalaryWorkTimeScreen';
 import { queryCompany } from '../../actions/timeAndSalaryCompany';
 
 const mapStateToProps = state => ({
@@ -14,4 +14,4 @@ const mapDispatchToProps = dispatch =>
 export default connect(
   mapStateToProps,
   mapDispatchToProps,
-)(TimeAndSalaryScreen);
+)(SalaryWorkTimeScreen);
