+++ src/containers/JobTitle/SalaryWorkTimeScreen.js
@@ -1,6 +1,6 @@
 import { connect } from 'react-redux';
 import { bindActionCreators } from 'redux';
-import TimeAndSalaryScreen from '../../components/JobTitle/TimeAndSalaryScreen';
+import SalaryWorkTimeScreen from '../../components/JobTitle/SalaryWorkTimeScreen';
 import { queryJobTitle } from '../../actions/timeAndSalaryJobTitle';
 
 const mapStateToProps = state => ({
@@ -14,4 +14,4 @@ const mapDispatchToProps = dispatch =>
 export default connect(
   mapStateToProps,
   mapDispatchToProps,
-)(TimeAndSalaryScreen);
+)(SalaryWorkTimeScreen);
