+++ src/routes.js
@@ -5,8 +5,8 @@ import LaborRightsSingle from './containers/LaborRightsSingle';
 import TimeAndSalary from './containers/SalaryWorkTime';
 import TimeAndSalaryBoard from './containers/SalaryWorkTime/TimeAndSalaryBoard';
 import SalaryWorkTimeSearchScreen from './containers/SalaryWorkTime/SearchScreen';
-import CompanyTimeAndSalaryScreen from './containers/Company/TimeAndSalaryScreen';
-import JobTitleTimeAndSalaryScreen from './containers/JobTitle/TimeAndSalaryScreen';
+import CompanySalaryWorkTimeScreen from './containers/Company/SalaryWorkTimeScreen';
+import JobTitleSalaryWorkTimeScreen from './containers/JobTitle/SalaryWorkTimeScreen';
 import TimeAndSalaryNotFound from './components/TimeAndSalary/NotFound';
 import CampaignTimeAndSalary from './containers/CampaignTimeAndSalary';
 import CampaignTimeAndSalaryBoard from './containers/CampaignTimeAndSalary/CampaignTimeAndSalaryBoard';
@@ -202,7 +202,7 @@ const routes = [
       {
         path: '/companies/:companyName/salary-work-times',
         exact: true,
-        component: CompanyTimeAndSalaryScreen,
+        component: CompanySalaryWorkTimeScreen,
       },
     ],
   },
@@ -214,7 +214,7 @@ const routes = [
       {
         path: '/job-titles/:jobTitle/salary-work-times',
         exact: true,
-        component: JobTitleTimeAndSalaryScreen,
+        component: JobTitleSalaryWorkTimeScreen,
       },
     ],
   },
